lens-backend/requirements 3.txt
fastapi>=0.109.0
mangum>=0.17.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.26.0
python-multipart>=0.0.6
